app/app/requirements.txt
Flask
mysql-connector
flask-mysqldb
flask-WTF